refactor: split TimelineView by concern
